+++ mypyc/test/test_emitmodule.py
@@ -10,14 +10,7 @@ from mypy import build
 from mypy.options import Options
 from mypyc.build import construct_groups
 from mypyc.codegen import emitmodule
-from mypyc.common import (
-    ENV_ATTR_NAME,
-    GENERATOR_ATTRIBUTE_PREFIX,
-    NEXT_LABEL_ATTR_NAME,
-    TEMP_ATTR_NAME,
-)
 from mypyc.errors import Errors
-from mypyc.ir.rtypes import RInstance
 from mypyc.irbuild.mapper import Mapper
 from mypyc.options import CompilerOptions
 
@@ -28,69 +21,6 @@ class FakeSCC:
 
 
 class TestEmitModule(unittest.TestCase):
-    def test_separate_generator_environment_keeps_private_frame_state(self) -> None:
-        with tempfile.TemporaryDirectory() as tmp_dir:
-            tmp_path = Path(tmp_dir)
-            module_path = tmp_path / "mod.py"
-            module_path.write_text(
-                """\
-from typing import Generator
-
-def make() -> str:
-    return "left"
-
-def outer() -> Generator[str, str, str]:
-    def nested(value: str) -> Generator[str, str, str]:
-        return make() + (yield value)
-    return nested("right")
-""",
-                encoding="utf-8",
-            )
-
-            sources = [build.BuildSource(str(module_path), "mod", None)]
-            options = Options()
-            options.preserve_asts = True
-            options.mypy_path = [str(tmp_path)]
-            options.cache_dir = str(tmp_path / ".mypy_cache")
-            options.per_module_options["mod"] = {"mypyc": True}
-
-            compiler_options = CompilerOptions(strict_traceback_checks=True)
-            groups = construct_groups(
-                sources, False, use_shared_lib=True, group_name_override=None
-            )
-            result = emitmodule.parse_and_typecheck(sources, options, compiler_options, groups)
-            try:
-                group_map = {
-                    source.module: lib_name for group, lib_name in groups for source in group
-                }
-                errors = Errors(options)
-                modules = emitmodule.compile_modules_to_ir(
-                    result, Mapper(group_map), compiler_options, errors
-                )
-                assert errors.num_errors == 0, errors.new_messages()
-            finally:
-                result.manager.metastore.close()
-
-            classes = [cl for module in modules.values() for cl in module.classes]
-            (generator,) = [cl for cl in classes if cl.has_running_flag]
-            assert generator.has_private_generator_frame
-            assert generator.attrs_are_thread_confined()
-
-            env_type = generator.attributes[ENV_ATTR_NAME]
-            assert isinstance(env_type, RInstance)
-            environment = env_type.class_ir
-            assert not environment.attrs_are_thread_confined()
-
-            private_attrs = set(generator.attributes)
-            shared_attrs = set(environment.attributes)
-            assert NEXT_LABEL_ATTR_NAME in private_attrs
-            assert NEXT_LABEL_ATTR_NAME in generator.attrs_with_defaults
-            assert NEXT_LABEL_ATTR_NAME not in shared_attrs
-            assert any(name.startswith(TEMP_ATTR_NAME) for name in private_attrs)
-            assert not any(name.startswith(TEMP_ATTR_NAME) for name in shared_attrs)
-            assert GENERATOR_ATTRIBUTE_PREFIX + "value" in shared_attrs
-            assert GENERATOR_ATTRIBUTE_PREFIX + "value" not in private_attrs
-
     def test_compile_modules_to_ir_orders_scc_members_deterministically(self) -> None:
         with tempfile.TemporaryDirectory() as tmp_dir, pytest.MonkeyPatch.context() as monkeypatch:
             tmp_path = Path(tmp_dir)
